+++ R/as_phyloseq.R
@@ -32,9 +32,11 @@
 #' @examples 
 #' \dontrun{
 #' # Install phyloseq to get example data
-#' # source('http://bioconductor.org/biocLite.R')
-#' # biocLite('phyloseq')
-#'
+#' # if (!require("BiocManager", quietly = TRUE))
+#' #   install.packages("BiocManager")
+#' # 
+#' # BiocManager::install("phyloseq")
+#' 
 #' # Parse example dataset
 #' library(phyloseq)
 #' data(GlobalPatterns)