+++ NEWS.md
@@ -11,7 +11,7 @@
 ##  metacoder 0.3.4
 
 * Added parsing of taxon ranks to the UNITE parser
-* Versious minor bug fixes
+* Various minor bug fixes
 
 ## metacoder 0.3.3