+++ NEWS.md
@@ -1,11 +1,12 @@
 # News 
 
-## Current
+## metacoder 0.3.6
 
 * When an ordered factor is supplied to the `groups` argument of `compare_groups`, the order of levels is used to arrange the results. This can be used to change the order of groups when the output is used with `heat_tree_matrix` (issue [#323](https://github.com/grunwaldlab/metacoder/issues/323)).
 * Added `calc_diff_abund_deseq2` function to use DESeq2 to produce output like `compare_group` useful for `heat_tree_matrix`
 * Fixed error with `primersearch` and `primersearch_raw` when there is a single match to a single query (issue [#326](https://github.com/grunwaldlab/metacoder/issues/326))
 * Fixed `Error in grid.Call.graphics(C_setviewport, vp, TRUE) : non-finite location and/or size for viewport` when using `heat_tree_matrix` with only a single comparison. Now a single differential heat tree is made instead of a matrix [#328](https://github.com/grunwaldlab/metacoder/issues/328)).
+* Replace use of depreciated `n_extra` in `print`.
 
 ## metacoder 0.3.5